progress to change oop code

+++ includes/delete_posts.php
@@ -5,8 +5,19 @@ session_start();
 include ("database_connection.php");
 
 
+if(isset($_GET['post_id'])){
+    $statement = $pdo->prepare("DELETE FROM posts WHERE post_id = :post_id");
+    // remove product from database
+    $statement->execute(
+        [
+            ":post_id" => $_GET["post_id"]
+        ]
+    );
+   
+}
 
-$statement = $pdo->prepare("DELETE FROM posts WHERE post_id = :post_id");
+
+/*$statement = $pdo->prepare("DELETE FROM posts WHERE post_id = :post_id");
 // remove product from database
 $statement->execute(
     [
@@ -15,5 +26,6 @@ $statement->execute(
 );
 $delete_posts = $statement->fetchAll(PDO::FETCH_ASSOC);
 // efter remove product, go to shoppingcart.php
+*/
 header('Location: ../views/admin_allposts.php');
 ?>
